# Sharpwaves performance fixes