@toni-neurosc 2024-01-26
Since Python for-loops are infamously slow, I vectorized as much as I could of the `analyze_waveform` function. This cut down the computation time for this method by almost another half, for `analyze_waveform` we're at 27% of the original computation time. 
_Rise steepness_ and _Decay steepness_ are a bit more complicated to vectorize so there's still a loop in there but I think I can use `scipy.ndimage` to get rid of it (not sure it would speed up anything but worth a try)

**HUGE DISCLAIMER**: I did my best to keep the math intact but I did this like 10 days ago so I'm not 100% sure the results are identical. Tests are passing but I haven't checked what they're doing.

It's also interesting that this opens the door to vectorizing the whole `calc_feature` function which might lead to further gains.

- After vectorizing (except steepness): 46.9043915271759 seconds per run.
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     2973    1.593    0.001   21.860    0.007 nm_sharpwaves.py:123(calc_feature)
    59460    5.236    0.000   13.390    0.000 nm_sharpwaves.py:240(analyze_waveform)
    29730    0.145    0.000    4.746    0.000 _signaltools.py:557(fftconvolve)
    29730    0.357    0.000    3.920    0.000 _signaltools.py:459(_freq_domain_conv)
   118920    0.533    0.000    3.720    0.000 _peak_finding.py:729(find_peaks)
```

- After vectorizing steepness: (TODO)