@toni-neurosc 2024-01-27
Fixed the test errors, caused by doing a [list] - [list] operation. I ran the tests in my computer and did not get the error, probably because self.sw_settings["sharpwave_features"]["width"] was not set. 
Do the tests in the GitHub Actions differ from the tests in /tests/?