@timonmerk 2024-01-29
Wow, that's great additions and impressive speed-up!
I will probably try to take some time to understand it write potentially additional tests that capture this feature.

Also, the tests in /tests are called by tox, so the Github Actions tests should be identical.