@toni-neurosc 2024-01-31
Update on steepness: I don't think this is possible to vectorize, due to the fact that each peak-to-trough / trough-to-peak distance is different. Using nd.image is possible but would require building masks out of the indexes, which requires extra computation, and using a view on numpy arrays as I have it currently is pretty efficient already anyway.

Btw, I wanted to ask, is there a possibility that there is a bug here:
```Python
rise_steepness = np.max(np.diff(self.data_process_sw[peak_idx_left : trough_idx + 1]))
decay_steepness = np.max(np.diff(self.data_process_sw[trough_idx : peak_idx_right + 1]) )
```

`np.diff` gives the difference to previous, so if peaks have higher values than troughs, rise_steepness will have mostly negative values, so should it not be np.min instead of max? 
Also I find it confusing that "rise" steepness is from peak to trough, feels like the naming should be reversed.