@timonmerk 2024-01-31
Yes! I think you're right. That's a bug. It should have been always the `np.abs()` before the `np.max()`. Good catch!